Rename MoveFoundsModal to MoveFundsModal (#4408)
* Rename MoveFoundsModal to MoveFundsModal

* fixup

* tests

---------
