+++ packages/ui/src/bounty/modals/AnnounceWorkEntryModal/AnnounceWorkEntryModal.tsx
@@ -9,7 +9,7 @@ import { useHasRequiredStake } from '@/accounts/hooks/useHasRequiredStake'
 import { useMyAccounts } from '@/accounts/hooks/useMyAccounts'
 import { useStakingAccountStatus } from '@/accounts/hooks/useStakingAccountStatus'
 import { useTransactionFee } from '@/accounts/hooks/useTransactionFee'
-import { MoveFundsModalCall } from '@/accounts/modals/MoveFoundsModal'
+import { MoveFundsModalCall } from '@/accounts/modals/MoveFundsModal'
 import { accountOrNamed } from '@/accounts/model/accountOrNamed'
 import { useApi } from '@/api/hooks/useApi'
 import { CurrencyName } from '@/app/constants/currency'
