+++ packages/ui/test/council/modals/AnnounceCandidacyModal.test.tsx
@@ -7,7 +7,7 @@ import React from 'react'
 import { Router } from 'react-router'
 import { interpret } from 'xstate'
 
-import { MoveFundsModalCall } from '@/accounts/modals/MoveFoundsModal'
+import { MoveFundsModalCall } from '@/accounts/modals/MoveFundsModal'
 import { ApiContext } from '@/api/providers/context'
 import { CurrencyName } from '@/app/constants/currency'
 import { GlobalModals } from '@/app/GlobalModals'
